abstraction in process
