+++ Wiwa/src/Wiwa/particles/ParticleManager.h
@@ -3,49 +3,124 @@
 #include <Wiwa/core/Core.h>
 #include <Wiwa/utilities/math/Vector3f.h>
 #include "Wiwa/ecs/components/ParticleEmitter.h"
+#include <Wiwa/utilities/render/TextureAnimation.h>
+#include <Wiwa/ecs/EntityManager.h>
 
 #include <vector>
 #include <string>
+#define MAX_PARTICLES 1000
+
 
 class Camera;
+class AnimationParticles;
+
 namespace Wiwa {
+
+
+	struct ParticleBillboard
+	{
+		Material* m_material;
+		glm::vec3	vertices[4];
+		glm::vec2	tex_coords[4];
+		int			vertex_indices[6];
+		Transform3D transform;
+
+		float		initial_lifetime = 0;
+		float		lifetime_percentage = 0;
+		float		lifetime = 0;
+		glm::vec4	color;
+
+		//position
+		glm::vec3	startingPosition;
+		glm::vec3	velocity;
+		glm::vec3	acceleration;
+
+		//rotation
+		glm::vec3	startingRotation;
+
+		//size
+		glm::vec3	start_size;
+		glm::vec3	growthVelocity;
+		glm::vec3	growthAcceleration;
+
+		bool		followEmitterRotation;
+		bool		followEmitterPosition;
+		bool		followParticle;
+
+		bool        isActive = false;
+
+		class AnimationParticles* animation = nullptr;
+
+		ParticleBillboard();
+
+	};
 
 	class WI_API ParticleManager
 	{
-	public:
-		struct EventData {
-			std::string name;
-			uint32_t hash;
-		};
-
-		struct BankData : public EventData {
-			std::string path;
-		};
-
-	private:
-		// Private constructor = default
-
-		// Set last error from result
-		static void setLastError(int result);
-
-		// =========== Variables ===========
 
 	public:
 
 		ParticleManager();
-		// Init particle engine
-		static bool Init();
+		~ParticleManager();
 
-		// Update particle events
-		static bool Update();
+		void Awake();
 
-		// Terminate particle engine
-		static bool Terminate();
+		void Init();
 
-		//To create a custom particle system
-		static bool CreateCustomParticle();
+		void Update();
 
-		static const uint32_t INVALID_ID = -1;
+		void OnDestroy();
+
+		void ScreenAlign(std::shared_ptr<ParticleBillboard> particle);
+
+		void EmitBatch();
+		void Play();
+		void Stop();
+
+	private:
+
+
+		ParticleEmitter* emitter;
+
+		EntityManager::ComponentIterator m_emitterComponent;
+
+		void DeleteParticleSystem();
+
+		void AddParticles();
+
+		void UpdateParticles();
+
+		float timer = 0;
+		float animationTimer = 0;
+		float dt;
+
+		int currentParticleIndex = 0;
+
+
+		std::array<ParticleBillboard, MAX_PARTICLES> particleArray;
+
+		class Material* m_Material;
+
+
+		glm::vec3 ref_vertices[4] =
+		{
+			glm::vec3(-1, 1, 0),	//0		0	  2   		1	  3
+			glm::vec3(-1, -1, 0),	//1				  		
+			glm::vec3(1, 1, 0),		//2				  		
+			glm::vec3(1, -1, 0)		//3		1	  3	  		0	  2
+
+		};
+
+		int ref_vertex_indices[6] =
+		{
+			0,1,3,
+			0,3,2
+		};
 	};
 
 }
+
+
+		
+
+
